Add ros_buildfarm_host_keys data bag.
This databag composes an item ID from the role and environment name and
queries for it specifically. This would allow the data bag to include
further items like repo.production or repo.development.

In order to work in situations with no default environment set you would
need to add a repo._default item.